+++ src/main/java/project/kingstagram/user/dto/response/UserProfileDTO.java
@@ -8,9 +8,9 @@ public class UserProfileDTO {
     private String userName;
     private String userNickname;
     private String userDescription;
-    private int postCount;
-    private int followerCount;
-    private int followingCount;
+    private Long postCount;
+    private Long followerCount;
+    private Long followingCount;
     private int responseCode;
     private String responseMessage;
 }
